[#24] Change the AI unit selection to account for effectiveness against a unit
The decision which unit of the units in range should be attacked, is now based on the damage the attacking unit can do on it and the distance to it.
So units will be favored when more damage can be done on them and when they're closer to the attacking unit.