[#55] Change the damage calculation to be no longer health-based
This was changed, because low-health units did almost no damage and players playing the game would not understand
what happened when they've attacked with a low-health unit. To keep the counter-attack and keep it being a lower damage attack,
every unit will do 100% damage when it's a normal attack and 50% when it's a counter-attack.